PythonProcess/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "PythonProcess"
authors = [{ name = "Jens Geudens" }]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
]

dependencies = ['grpcio==1.76.0', 'grpcio-tools==1.76.0', 'protobuf==6.33.4']

dynamic = ["version"]

[project.optional-dependencies]
development = [
    # Quality dependencies
    'black==25.12.0',
    'pydocstyle==6.3.0',
    'codespell==2.4.1',
    'mypy==1.19.1',
    'flake8==7.3.0',
    'isort==7.0.0',
    'pylint==4.0.4',
    'pre-commit==4.5.1',
    'types-protobuf==6.32.1.20251210',
    'types-grpcio==1.0.0.20251009',
    'mypy-protobuf==5.0.0',
]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120
max-complexity = 10
select = "B,C,E,F,W,T4"
ignore = "W503"

[tool.mypy]
files = "src"
ignore_missing_imports = true
strict = true

[tool.doc8]
max-line-length = 120
verbose = 0
ignore = "D002,D004"

[tool.pylint.format]
max-line-length = 120
disable = "logging-fstring-interpolation"
min-similarity-lines = 10

[tool.pydocstyle]
convention = "google"

[tool.pytest]

[tool.codespell]
skip = "./venv,.mypy_cache,./build,*.patch,.git, data/*"
